Chat Renewed
After a few months of working on this chat here and there. One feature
kept leading to another and before I knew it, there was way too much to
commit over. In fact, this version isn't compatible with the legacy
chat.

Now for changes, this is what I can remember off the top of my head

• Reworked message sending (now sends as an array)
• Added mute command
• Added message deleting (admins can click on timestamp to delete a
message)
• Added Function comments
• Added momentjs (along with user timezone)
• Added time stamps to all cmd messages
• Made it so the chat can be 'worked on' w/o an internet connection
• Cleaned code (still needs a little more work)
• Updated README
• Removed node_modules from the repository

I still have a lot of stuff in mind I still want to add, so stay tuned
:)