Emotes & Images Update
So the main features of this update are custom emotes and images.
They're more or less in a beta stage since it's not perfect. There's
also an issue of images loading after the scroll bar updates, not sure
how to handle that yet.

This update also includes some code cleaning and fixes.